Guard copy of artifacts on existence
- Hopefully, this stop the situation where a partial build still exports
artifacts, and breaks the 'latest' build symlink...

+++ kokoro/continuous.sh
@@ -35,10 +35,16 @@ python3 ${ROOTDIR}/build/create_release_manifest.py \
   -o ${KOKORO_ARTIFACTS_DIR}/manifest.xml
 popd
 
-cp ${ROOTDIR}/board/flash.sh ${KOKORO_ARTIFACTS_DIR}
-chmod -x ${KOKORO_ARTIFACTS_DIR}/flash.sh
-cp ${PRODUCT_OUT}/u-boot.imx ${KOKORO_ARTIFACTS_DIR}
-cp ${PRODUCT_OUT}/boot.img ${KOKORO_ARTIFACTS_DIR}
-cp ${PRODUCT_OUT}/partition-table-*.img ${KOKORO_ARTIFACTS_DIR}
-cp ${PRODUCT_OUT}/rootfs.img ${KOKORO_ARTIFACTS_DIR}
-cp ${PRODUCT_OUT}/sdcard.img ${KOKORO_ARTIFACTS_DIR}
+if [[ -f ${PRODUCT_OUT}/u-boot.imx && \
+      -f ${PRODUCT_OUT}/boot.img && \
+      -f ${PRODUCT_OUT}/partition-table-8gb.img && \
+      -f ${PRODUCT_OUT}/rootfs.img && \
+      -f ${PRODUCT_OUT}/sdcard.img ]]; then
+  cp ${ROOTDIR}/board/flash.sh ${KOKORO_ARTIFACTS_DIR}
+  chmod -x ${KOKORO_ARTIFACTS_DIR}/flash.sh
+  cp ${PRODUCT_OUT}/u-boot.imx ${KOKORO_ARTIFACTS_DIR}
+  cp ${PRODUCT_OUT}/boot.img ${KOKORO_ARTIFACTS_DIR}
+  cp ${PRODUCT_OUT}/partition-table-*.img ${KOKORO_ARTIFACTS_DIR}
+  cp ${PRODUCT_OUT}/rootfs.img ${KOKORO_ARTIFACTS_DIR}
+  cp ${PRODUCT_OUT}/sdcard.img ${KOKORO_ARTIFACTS_DIR}
+fi
